+++ packages/runtime-core/src/renderer.ts
@@ -65,6 +65,7 @@ import { createHydrationFunctions, RootHydrateFunction } from './hydration'
 import { invokeDirectiveHook } from './directives'
 import { startMeasure, endMeasure } from './profiling'
 import { ComponentPublicInstance } from './componentProxy'
+import { componentRemoved, componentUpdated } from './devtools'
 
 export interface Renderer<HostElement = RendererElement> {
   render: RootRenderFunction<HostElement>
@@ -638,7 +639,31 @@ function baseCreateRenderer(
         optimized
       )
     } else {
-      patchElement(n1, n2, parentComponent, parentSuspense, isSVG, optimized)
+      if (
+        __DEV__ &&
+        isHmrUpdating &&
+        hostCloneNode !== undefined &&
+        n2.patchFlag === PatchFlags.HOISTED
+      ) {
+        // https://github.com/vitejs/vite/issues/514
+        // reused hoisted trees are inserted with cloneNode
+        // which makes them not patch-able. In production hoisted trees are
+        // never patched (because they are not collected as dynamic nodes), but
+        // they can be udpated during HMR. In this case just mount it as new
+        // and remove the stale DOM tree.
+        mountElement(
+          n2,
+          container,
+          n1.el,
+          parentComponent,
+          parentSuspense,
+          isSVG,
+          optimized
+        )
+        hostRemove(n1.el!)
+      } else {
+        patchElement(n1, n2, parentComponent, parentSuspense, isSVG, optimized)
+      }
     }
   }
 
@@ -736,14 +761,17 @@ function baseCreateRenderer(
     }
 
     hostInsert(el, container, anchor)
+    // #1583 For inside suspense case, enter hook should call when suspense resolved
+    const needCallTransitionHooks =
+      !parentSuspense && transition && !transition.persisted
     if (
       (vnodeHook = props && props.onVnodeMounted) ||
-      (transition && !transition.persisted) ||
+      needCallTransitionHooks ||
       dirs
     ) {
       queuePostRenderEffect(() => {
         vnodeHook && invokeVNodeHook(vnodeHook, parentComponent, vnode)
-        transition && !transition.persisted && transition.enter(el)
+        needCallTransitionHooks && transition!.enter(el)
         dirs && invokeDirectiveHook(vnode, null, parentComponent, 'mounted')
       }, parentSuspense)
     }
@@ -1390,6 +1418,7 @@ function baseCreateRenderer(
         }
         if (__DEV__) {
           popWarningContext()
+          componentUpdated(instance)
         }
       }
     }, __DEV__ ? createDevEffectOptions(instance) : prodEffectOptions)
@@ -2041,6 +2070,8 @@ function baseCreateRenderer(
         parentSuspense.resolve()
       }
     }
+
+    __DEV__ && componentRemoved(instance)
   }
 
   const unmountChildren: UnmountChildrenFn = (