+++ test-dts/reactivity.test-d.ts
@@ -1,0 +1,9 @@
+import { readonly, describe, expectError } from './index'
+
+describe('should support DeepReadonly', () => {
+  const r = readonly({ obj: { k: 'v' } })
+  // @ts-expect-error
+  expectError((r.obj = {}))
+  // @ts-expect-error
+  expectError((r.obj.k = 'x'))
+})